+++ nodejs/examples/01_the_switch_and_the_light/light.js
@@ -42,7 +42,7 @@ client.on('update', (event) => {
       throw err;
     }
   });
-  console.log('New desired state is', desired_state ? 'ON' : 'OFF');
+  console.log('Light is now', desired_state ? 'ON' : 'OFF');
 });
 
 // Register shutdown handling