+++ rust/src/cache.rs
@@ -1,9 +1,7 @@
-use super::{Stats};
+use super::Stats;
 use serde::Deserialize;
 use serde_json::Value;
-use std::{
-    collections::HashMap,
-};
+use std::collections::HashMap;
 
 #[derive(Clone, Debug, Default, Deserialize)]
 pub struct Cache {