+++ rust/src/lib.rs
@@ -1,11 +1,11 @@
+mod cache;
 mod client;
 mod error;
 mod stats;
 mod system;
-mod cache;
 
-pub use client::Client;
 pub use cache::Cache;
+pub use client::Client;
 pub use error::Error;
 pub use stats::Stats;
 use std::sync::{Arc, Mutex};