+++ rust/examples/01_the_switch_and_the_light/light.rs
@@ -43,9 +43,9 @@ pub fn main() {
     client.add_context(NODE_ID, CONTEXT_ID, CONTEXT_NAME).unwrap();
     eprintln!("Registered context {CONTEXT_ID} for node {NODE_ID} on Arete control plane");
 
-    // Register the "padi.light" profile with the context
-    client.add_profile(NODE_ID, CONTEXT_ID, PADI_LIGHT_PROFILE);
-    eprintln!("Registered profile {PADI_LIGHT_PROFILE} for context {CONTEXT_ID} on Arete control plane");
+    // Register as a consumer of state for the "padi.light" profile
+    client.add_consumer(NODE_ID, CONTEXT_ID, PADI_LIGHT_PROFILE);
+    eprintln!("Registered as consumer of state for {PADI_LIGHT_PROFILE} profile for context {CONTEXT_ID} on Arete control plane");
 
     // Realize initial desired state
     if let Some(value) = client.get(DESIRED_STATE_KEY, Some("0".into())).unwrap() {