+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Project Arete SDK Changelog
 
+## [Unreleased]
+### Changed
+- [#54](https://github.com/project-arete/sdk/issues/54) (Rust) Upgrade Rust 1.86.0 → 1.88.0
+
 ## [0.1.3] - 2025-09-05
 ### Added
 - [#46](https://github.com/project-arete/sdk/issues/46) Example services should add their system to the Arete Control Plane on startup