+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Changed
 - [#54](https://github.com/project-arete/sdk/issues/54) (Rust) Upgrade Rust 1.86.0 → 1.88.0
 
+### Fixed
+- [#51](https://github.com/project-arete/sdk/issues/51) (Python) put is not implemented
+
 ## [0.1.3] - 2025-09-05
 ### Added
 - [#46](https://github.com/project-arete/sdk/issues/46) Example services should add their system to the Arete Control Plane on startup