+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Project Arete SDK Changelog
 
-## [Unreleased]
+## [0.1.4] - 2025-09-12
 ### Added
 - [#71](https://github.com/project-arete/sdk/issues/71) Expose Client addProvider and addConsumer fns
 - [#69](https://github.com/project-arete/sdk/issues/69) Example services should register the "padi.light" profile with their context on the Arete Control Plane on startup